Reinstated real(dp) back into all fortran source files

+++ source/fortran/availability.f90
@@ -19,10 +19,10 @@ module availability_module
   ! Module subroutine and variable declarations !
   ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-  real(8), parameter :: year = 31557600.0D0
+  real(dp), parameter :: year = 31557600.0D0
   !! seconds in a year [s]
 
-  real(8), parameter :: day = 86400.0D0
+  real(dp), parameter :: day = 86400.0D0
   !! seconds in a day [s]
 
 contains
@@ -60,11 +60,11 @@ contains
     !  Local variables !
     ! !!!!!!!!!!!!!!!!!!!
 
-    real(8) :: lb, ld, td
-    real(8) :: uplanned, uutot
+    real(dp) :: lb, ld, td
+    real(dp) :: uplanned, uutot
     integer :: n
 
-    real(8) :: n_sec_year
+    real(dp) :: n_sec_year
     !! Number of seconds in a year
     
     ! Full power lifetime (in years)
@@ -237,14 +237,14 @@ contains
     !  Local variables !
     ! !!!!!!!!!!!!!!!!!!!
 
-    real(8) :: u_planned
-    real(8) :: u_unplanned
-    real(8) :: u_unplanned_magnets
-    real(8) :: u_unplanned_div
-    real(8) :: u_unplanned_fwbs
-    real(8) :: u_unplanned_bop
-    real(8) :: u_unplanned_hcd
-    real(8) :: u_unplanned_vacuum
+    real(dp) :: u_planned
+    real(dp) :: u_unplanned
+    real(dp) :: u_unplanned_magnets
+    real(dp) :: u_unplanned_div
+    real(dp) :: u_unplanned_fwbs
+    real(dp) :: u_unplanned_bop
+    real(dp) :: u_unplanned_hcd
+    real(dp) :: u_unplanned_vacuum
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -336,15 +336,15 @@ contains
 
     ! Arguments
     integer, intent(in) :: outfile, iprint
-    real(8), intent(out) :: u_planned
+    real(dp), intent(out) :: u_planned
 
     ! Local variables !
     ! !!!!!!!!!!!!!!!!!!
-    real(8) :: n_sec_year
+    real(dp) :: n_sec_year
     !! Number of seconds in a year
     
-    real(8) :: mttr_blanket, mttr_divertor, mttr_shortest
-    real(8) :: lifetime_shortest, lifetime_longest
+    real(dp) :: mttr_blanket, mttr_divertor, mttr_shortest
+    real(dp) :: lifetime_shortest, lifetime_longest
     integer :: n
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -454,14 +454,14 @@ contains
 
     ! Arguments
     integer, intent(in) :: outfile, iprint
-    real(8), intent(out) :: u_unplanned_magnets
+    real(dp), intent(out) :: u_unplanned_magnets
 
     ! Local Variables !
     ! !!!!!!!!!!!!!!!!!!
 
-    real(8) :: mag_temp_marg_limit, mag_temp_marg, mag_main_time
-    real(8) :: mag_min_u_unplanned, start_of_risk, t_life
-    real(8) :: tmargmin
+    real(dp) :: mag_temp_marg_limit, mag_temp_marg, mag_main_time
+    real(dp) :: mag_min_u_unplanned, start_of_risk, t_life
+    real(dp) :: tmargmin
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -531,12 +531,12 @@ contains
 
     ! Arguments
     integer, intent(in) :: outfile, iprint
-    real(8), intent(out) :: u_unplanned_div
+    real(dp), intent(out) :: u_unplanned_div
 
     ! Local Variables !
     ! !!!!!!!!!!!!!!!!!!
 
-    real(8) :: a0, div_avail, n, pf
+    real(dp) :: a0, div_avail, n, pf
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -613,12 +613,12 @@ contains
 
     ! Arguments
     integer, intent(in) :: outfile, iprint
-    real(8), intent(out) :: u_unplanned_fwbs
+    real(dp), intent(out) :: u_unplanned_fwbs
 
     ! Local Variables !
     ! !!!!!!!!!!!!!!!!!!
 
-    real(8) :: a0, fwbs_avail, n, pf
+    real(dp) :: a0, fwbs_avail, n, pf
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -691,12 +691,12 @@ contains
 
     ! Arguments
     integer, intent(in) :: outfile, iprint
-    real(8), intent(out) :: u_unplanned_bop
+    real(dp), intent(out) :: u_unplanned_bop
 
     ! Local variables !
     ! !!!!!!!!!!!!!!!!!!
 
-    real(8) :: bop_fail_rate, bop_mttr
+    real(dp) :: bop_fail_rate, bop_mttr
     integer :: bop_num_failures
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -749,7 +749,7 @@ contains
     ! Arguments !
     ! !!!!!!!!!!!!
 
-    real(8), intent(out) :: u_unplanned_hcd
+    real(dp), intent(out) :: u_unplanned_hcd
 
     ! Currently just a fixed value until more information available or Q.
     ! Tran's response provides useful data.
@@ -782,15 +782,15 @@ contains
 
     ! Arguments
     integer, intent(in) :: outfile, iprint
-    real(8), intent(out) :: u_unplanned_vacuum
+    real(dp), intent(out) :: u_unplanned_vacuum
 
     ! Local variables
     integer :: total_pumps, n
-    real(8) :: cryo_failure_rate, cryo_main_time
-    real(8) :: cryo_nfailure_rate, t_down
-    real(8) :: n_shutdown, t_op_bt, sum_prob
+    real(dp) :: cryo_failure_rate, cryo_main_time
+    real(dp) :: cryo_nfailure_rate, t_down
+    real(dp) :: n_shutdown, t_op_bt, sum_prob
 
-    real(8), dimension(vpumpn + redun_vac + 1) :: vac_fail_p
+    real(dp), dimension(vpumpn + redun_vac + 1) :: vac_fail_p
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
